`fn rav1d_num_logical_processors`: Make return type a `usize`.

+++ src/lib.rs
@@ -218,7 +218,7 @@ unsafe fn get_num_threads(
     *n_tc = (if (*s).n_threads != 0 {
         (*s).n_threads
     } else {
-        iclip(rav1d_num_logical_processors(), 1 as c_int, 256 as c_int)
+        iclip(rav1d_num_logical_processors() as c_int, 1 as c_int, 256 as c_int)
     }) as c_uint;
     *n_fc = if (*s).max_frame_delay != 0 {
         cmp::min((*s).max_frame_delay as c_uint, *n_tc)
